restructured commands
